requirements.txt
decouple
python-telegram-bot
azure-ai-inference